+++ Mindspace_back/src/node/dto/node-info-response.dto.ts
@@ -2,10 +2,10 @@ import { ApiProperty } from '@nestjs/swagger';
 import { NodeLinkDto } from './node-link.dto';
 
 export class NodeInfoResponse {
-    @ApiProperty({ type: [Object] })
-    // nodes: Map<string, Object>[];
-    nodes: any[];
+  @ApiProperty({ type: [Object] })
+  // nodes: Map<string, Object>[];
+  nodes: any[];
 
-    @ApiProperty({ type: [NodeLinkDto] })
-    links: NodeLinkDto[];
+  @ApiProperty({ type: [NodeLinkDto] })
+  links: NodeLinkDto[];
 }
