+++ Mindspace_back/src/node/dto/node-link.dto.ts
@@ -1,9 +1,9 @@
 import { ApiProperty } from '@nestjs/swagger';
 
 export class NodeLinkDto {
-    @ApiProperty({ description: 'source', example: 'source11' })
-    source: number;
+  @ApiProperty({ description: 'source', example: 'source11' })
+  source: number;
 
-    @ApiProperty({ description: 'target', example: 'target11' })
-    target: number;
+  @ApiProperty({ description: 'target', example: 'target11' })
+  target: number;
 }
