+++ Mindspace_back/src/node/entities/node.entity.ts
@@ -3,11 +3,9 @@ import { Timestamp } from '../../global/common/timeStamp';
 
 @Entity()
 export class Node extends Timestamp {
+  @PrimaryGeneratedColumn('increment', { name: 'node_id' })
+  id: number;
 
-    @PrimaryGeneratedColumn('increment', { name: 'node_id' })
-    id: number;
-
-    @Column({ name: 'node_name', type: 'varchar', nullable: false })
-    name: string;
-
+  @Column({ name: 'node_name', type: 'varchar', nullable: false })
+  name: string;
 }
