+++ Mindspace_back/src/node/node.controller.ts
@@ -2,8 +2,8 @@ import { Controller, Get, Headers } from '@nestjs/common';
 import { NodeService } from './node.service';
 import { Node } from './entities/node.entity';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
-import {NodeInfoResponse} from "./dto/node-info-response.dto";
-import {NodeResponseDto} from "./dto/node-response.dto";
+import { NodeInfoResponse } from './dto/node-info-response.dto';
+import { NodeResponseDto } from './dto/node-response.dto';
 
 @ApiTags('Node')
 @Controller('api/v1/node')
@@ -17,9 +17,10 @@ export class NodeController {
   }
 
   @Get('check')
-  async getNodeInfoWithLinks(@Headers('authorization') userId: number): Promise<NodeInfoResponse> {
+  async getNodeInfoWithLinks(
+    @Headers('authorization') userId: number,
+  ): Promise<NodeInfoResponse> {
     const nodeInfo = await this.nodeService.getNodeInfoWithLinks(userId);
     return nodeInfo;
   }
-
 }
